+++ pyrefly/lib/alt/call.rs
@@ -12,8 +12,8 @@ use starlark_map::small_set::SmallSet;
 use vec1::Vec1;
 use vec1::vec1;
 
-use crate::alt::answers::AnswersSolver;
 use crate::alt::answers::LookupAnswer;
+use crate::alt::answers_solver::AnswersSolver;
 use crate::alt::attr::DescriptorBase;
 use crate::alt::callable::CallArg;
 use crate::alt::callable::CallKeyword;
