+++ pyrefly/lib/alt/class/classdef.rs
@@ -13,8 +13,8 @@ use ruff_python_ast::StmtClassDef;
 use ruff_python_ast::name::Name;
 use starlark_map::small_map::SmallMap;
 
-use crate::alt::answers::AnswersSolver;
 use crate::alt::answers::LookupAnswer;
+use crate::alt::answers_solver::AnswersSolver;
 use crate::alt::class::class_field::ClassField;
 use crate::alt::types::class_metadata::ClassMetadata;
 use crate::alt::types::class_metadata::ClassMro;
