+++ pyrefly/lib/alt/class/named_tuple.rs
@@ -9,8 +9,8 @@ use ruff_python_ast::name::Name;
 use starlark_map::small_set::SmallSet;
 use starlark_map::smallmap;
 
-use crate::alt::answers::AnswersSolver;
 use crate::alt::answers::LookupAnswer;
+use crate::alt::answers_solver::AnswersSolver;
 use crate::alt::types::class_metadata::ClassSynthesizedField;
 use crate::alt::types::class_metadata::ClassSynthesizedFields;
 use crate::error;
