+++ pyrefly/lib/alt/class/typed_dict.rs
@@ -16,8 +16,8 @@ use starlark_map::small_set::SmallSet;
 use starlark_map::smallmap;
 use vec1::Vec1;
 
-use crate::alt::answers::AnswersSolver;
 use crate::alt::answers::LookupAnswer;
+use crate::alt::answers_solver::AnswersSolver;
 use crate::alt::types::class_metadata::ClassMetadata;
 use crate::alt::types::class_metadata::ClassSynthesizedField;
 use crate::alt::types::class_metadata::ClassSynthesizedFields;
