+++ pyrefly/lib/alt/mod.rs
@@ -6,6 +6,7 @@
  */
 
 pub mod answers;
+pub mod answers_solver;
 pub mod attr;
 pub mod call;
 pub mod callable;
