+++ pyrefly/lib/alt/narrow.rs
@@ -20,8 +20,8 @@ use ruff_python_ast::name::Name;
 use ruff_text_size::Ranged;
 use ruff_text_size::TextRange;
 
-use crate::alt::answers::AnswersSolver;
 use crate::alt::answers::LookupAnswer;
+use crate::alt::answers_solver::AnswersSolver;
 use crate::alt::attr::Narrowable;
 use crate::alt::callable::CallArg;
 use crate::alt::callable::CallKeyword;
