+++ pyrefly/lib/state/state.rs
@@ -61,13 +61,13 @@ use vec1::vec1;
 use crate::alt::answers::AnswerEntry;
 use crate::alt::answers::AnswerTable;
 use crate::alt::answers::Answers;
-use crate::alt::answers::AnswersSolver;
-use crate::alt::answers::CalcStack;
-use crate::alt::answers::Cycles;
 use crate::alt::answers::LookupAnswer;
 use crate::alt::answers::Solutions;
 use crate::alt::answers::SolutionsEntry;
 use crate::alt::answers::SolutionsTable;
+use crate::alt::answers_solver::AnswersSolver;
+use crate::alt::answers_solver::CalcStack;
+use crate::alt::answers_solver::Cycles;
 use crate::alt::traits::Solve;
 use crate::binding::binding::Exported;
 use crate::binding::binding::KeyExport;
